+++ src/PlayfieldLayout.h
@@ -14,21 +14,21 @@ namespace PlayfieldLayout
 {
 
 // Centered vertically, with the paddle's left edge inset by `wallInset` logical pixels from the left wall.
-SDL_FRect leftPaddle(int playfieldWidth,
-                     int playfieldHeight,
+SDL_FRect leftPaddle(float playfieldWidth,
+                     float playfieldHeight,
                      float paddleHalfWidth,
                      float paddleHalfHeight,
                      float wallInset);
 
 // Mirror of leftPaddle on the right wall (same vertical centering, same inset measured from the right wall).
-SDL_FRect rightPaddle(int playfieldWidth,
-                      int playfieldHeight,
+SDL_FRect rightPaddle(float playfieldWidth,
+                      float playfieldHeight,
                       float paddleHalfWidth,
                       float paddleHalfHeight,
                       float wallInset);
 
 // Ball centered on both axes of the playfield.
-SDL_FRect ball(int playfieldWidth, int playfieldHeight, float ballHalfSize);
+SDL_FRect ball(float playfieldWidth, float playfieldHeight, float ballHalfSize);
 
 // Evenly distributed dashes for the vertical center line.
 //
@@ -45,8 +45,8 @@ SDL_FRect ball(int playfieldWidth, int playfieldHeight, float ballHalfSize);
 //
 // The `gap` parameter is the documented vertical spacing *between* consecutive dashes; the caller picks it (typically
 // from Playfield::kCenterDashGap) so the helper does not need to know the production tuning.
-std::vector<SDL_FRect> centerDashSegments(int playfieldWidth,
-                                          int playfieldHeight,
+std::vector<SDL_FRect> centerDashSegments(float playfieldWidth,
+                                          float playfieldHeight,
                                           int segmentCount,
                                           float dashWidth,
                                           float dashHeight,
